Add font unpacking and repacking
- For some reason DSFont24.tpf.dcx seems to be completely ignored
by the game. Since this is the font used in most situations, this
makes translations to langauges with special characters very
problematic. Extra note: Unlike talkfont's, dsFont's filesize
decreases drastically (300KB, ~1/6 of full size) during the first
un/repacking, the cause of which is unknown.
- IMPORTANT: When editing .dds (texture) files, use these settings
when saving to keep the original format of the file:
	- mode: BC3 (linear, DXT5)
	- Mip maps on
	- nearest neighbour